requirements.txt
click==8.0.4
dlib==19.23.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.22.3
opencv-python==4.5.5.62
Pillow==9.0.1

tensorflow~=2.8.0
mtcnn~=0.1.1